WIP - Path handling cleanup and integrating latest metabase

+++ metabuilder/src/common.cpp
@@ -405,7 +405,7 @@ static int luaFuncCheckPlatform(lua_State* l)
 
     std::string testPlatform;
 	mbLuaToStringExpandMacros(&testPlatform, b, l, 1);
-	for (int i = 0; i < (int)mbGetActiveContext()->metabase->supportedPlatforms.size(); ++i)
+	for (size_t i = 0; i < mbGetActiveContext()->metabase->supportedPlatforms.size(); ++i)
 	{
 		const std::string& test = mbGetActiveContext()->metabase->supportedPlatforms[i];
 		if (test == testPlatform)
@@ -512,7 +512,7 @@ void mbHostPathJoin(char* result, const char* a, const char* b)
 	}
 
 	//Trim trailing slash
-	int aLen = (int)strlen(a);
+	int aLen = static_cast<int>(strlen(a));
 
 	char trailingSlashToRestore = 0;
 	char* trailingSlashPtr = NULL;
@@ -534,7 +534,7 @@ void mbHostPathJoin(char* result, const char* a, const char* b)
 std::string mbPathGetDir(const std::string& filePath)
 {
     std::string tmp;
-    int len = (int)filePath.length();
+    int len = static_cast<int>(filePath.length());
     char* chars = (char*)filePath.c_str();
     for (int i = len-1; i >= 0; --i)
     {
@@ -554,7 +554,7 @@ std::string mbPathGetDir(const std::string& filePath)
 std::string	mbPathGetFilename(const std::string& filePath)
 {
     std::string tmp;
-    int len = (int)filePath.length();
+    int len = static_cast<int>(filePath.length());
     char* chars = (char*)filePath.c_str();
     for (int i = len-1; i >= 0; --i)
     {
@@ -570,7 +570,7 @@ std::string	mbPathGetFilename(const std::string& filePath)
 
 bool mbPathGetFileExtension(char* result, const char* filename)
 {
-    int len = (int)strlen(filename);
+    int len = static_cast<int>(strlen(filename));
     const char* chars = filename;
     for (int i = len-1; i >= 0; --i)
     {
@@ -589,7 +589,7 @@ bool mbPathReplaceFileExtension(char* result, const char* filename, const char* 
 {
 	strcpy(result, filename);
 	
-    int len = (int)strlen(result);
+    int len = static_cast<int>(strlen(result));
 	char* chars = result;
     for (int i = len-1; i >= 0; --i)
     {
@@ -701,8 +701,8 @@ void mbPopDir()
 
 int luaCreateTable(lua_State* l)
 {
-	int narr = (int)lua_tonumber(l, 1);
-	int nrec = (int)lua_tonumber(l, 2);
+	int narr = static_cast<int>(lua_tonumber(l, 1));
+	int nrec = static_cast<int>(lua_tonumber(l, 2));
 
 	lua_createtable(l, narr, nrec);
 
@@ -806,8 +806,8 @@ void mbMergeOptions(std::map<std::string, KeyValueMap>* result,	const std::map<s
 
 U32 mbRandomU32(mbRandomContext& ctx)
 {
-// Originally by David Jones, UCL in http://www.cs.ucl.ac.uk/staff/d.jones/GoodPracticeRNG.pdf
-// Public domain code for JKISS RNG
+    // Originally by David Jones, UCL in http://www.cs.ucl.ac.uk/staff/d.jones/GoodPracticeRNG.pdf
+    // Public domain code for JKISS RNG
 	U32& x = ctx.seed[0];
 	U32& y = ctx.seed[1];
 	U32& z = ctx.seed[2];
@@ -841,7 +841,7 @@ void mbCheckExpectedBlock(E_BlockType blockExpected, const char* cmdName)
 void mbJoinArrays(StringVector* a, const StringVector& b)
 {
 	a->reserve(a->size() + b.size());
-	for (int i = 0; i < (int)b.size(); ++i)
+	for (size_t i = 0; i < b.size(); ++i)
 	{
 		a->push_back(b[i]);
 	}
@@ -861,7 +861,7 @@ void mbRemoveDuplicates(StringVector* strings_)
 	tmp.reserve(strings.size());
 	
 	UniqueStringHashTable uniqueStrings;
-	for (int i = 0; i < (int)strings.size(); ++i)
+	for (size_t i = 0; i < strings.size(); ++i)
 	{
 		UniqueStringHashTable::const_iterator it = uniqueStrings.find(&strings[i]);
 		if (it == uniqueStrings.end())
@@ -896,7 +896,7 @@ void mbRemoveDuplicatesAndSort(StringVector* strings_)
 	tmp.reserve(strings.size());
 		
 	UniqueStringHashTable uniqueStrings;
-	for (int i = 0; i < (int)strings.size(); ++i)
+	for (size_t i = 0; i < strings.size(); ++i)
 	{
 		uniqueStrings.insert(&strings[i]);
 	}
@@ -915,7 +915,7 @@ void mbRemoveDuplicatesAndSort(StringVector* strings_)
 	std::sort(tmp.begin(), tmp.end(), mbCompareNoCase);
 
 	strings.clear();
-	for (int i = 0; i < (int)tmp.size(); ++i)
+	for (size_t i = 0; i < tmp.size(); ++i)
 	{
 		StringSortRecord* r = tmp[i];
 		strings.push_back(r->originalString);
@@ -991,7 +991,7 @@ static int mbExpandSingleMacro(char* result, const KeyValueMap& macroMap, const 
 	{
 		const std::string& value = it->second;
 		memcpy(result, value.c_str(), value.length());
-		length = (int)value.length();
+		length = static_cast<int>(value.length());
 	}
 
 	if (length == 0)
@@ -999,7 +999,7 @@ static int mbExpandSingleMacro(char* result, const KeyValueMap& macroMap, const 
 		const char* envValue = getenv(macroKey);
 		if (envValue)
 		{
-			length = (int)strlen(envValue);
+			length = static_cast<int>(strlen(envValue));
 			memcpy(result, envValue, length);
 		}
 	}
